backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
SQLAlchemy
psycopg2-binary
passlib[bcrypt]
python-jose
python-dotenv
